+++ tests/test_registerable.py
@@ -13,6 +13,7 @@ from tests.test_utils import (
     authenticate,
     check_xlation,
     get_form_input,
+    init_app_with_options,
     json_authenticate,
     logout,
 )
@@ -931,3 +932,27 @@ def test_subclass(app, sqlalchemy_datastore):
     )
     assert response.status_code == 400
     assert "Really - don't start" in response.json["response"]["errors"][0]
+
+
+@pytest.mark.confirmable()
+@pytest.mark.settings(return_generic_responses=True)
+def test_my_mail_util(app, sqlalchemy_datastore):
+    # Test that with generic_responses - we still can get syntax/validation errors.
+    from flask_security import MailUtil, EmailValidateException
+
+    class MyMailUtil(MailUtil):
+        def validate(self, email):
+            if email.startswith("mike"):
+                raise EmailValidateException("No mikes allowed")
+
+    init_app_with_options(
+        app, sqlalchemy_datastore, **{"security_args": {"mail_util_cls": MyMailUtil}}
+    )
+
+    data = dict(
+        email="mike@lp.com",
+        password="awesome sunset",
+    )
+    client = app.test_client()
+    response = client.post("/register", data=data)
+    assert b"No mikes allowed" in response.data